Add validation for user input in AddPerson and EditPerson

+++ web/src/main/java/com/exercise/servlet1/ManageRoles.java
@@ -7,12 +7,16 @@ import javax.servlet.ServletException;
 import javax.servlet.http.HttpServlet;
 import javax.servlet.http.HttpServletRequest;
 import javax.servlet.http.HttpServletResponse;
+
+import org.hibernate.exception.ConstraintViolationException;
+
 import com.exercise.servlet1.core.*;
 
 public class ManageRoles extends HttpServlet{
 	
 	private static final long serialVersionUID = 1L;
 	private RoleService roleService = new RoleService();
+	private PersonService personService = new PersonService();
 	private Validation check = new Validation();
 	
 	public ManageRoles(){
@@ -28,12 +32,7 @@ public class ManageRoles extends HttpServlet{
 	}
 	
 	protected void doGet(HttpServletRequest request, HttpServletResponse response) throws ServletException, IOException{
-		System.out.println("ListRoles doGet() method called");
-		System.out.println("Action = View Roles");
-			
 		List<Role> roles = roleService.getRoles();
-		response.setContentType("text/html");
-		response.setStatus(HttpServletResponse.SC_OK);
 		request.setAttribute("roles", roles);
 		request.getRequestDispatcher("/WEB-INF/ViewRoles.jsp").forward(request, response);
 	}
@@ -41,15 +40,15 @@ public class ManageRoles extends HttpServlet{
 	protected void doPost(HttpServletRequest request, HttpServletResponse response) throws ServletException, IOException{
 		String action = request.getParameter("action");
 		if(action.equals("addRole")){
-			String message=null;
 			String newRole = request.getParameter("newRole");
-				if(newRole.isEmpty()){
-					message = "You entered an empty field";
-					request.setAttribute("errors", message);
+				newRole = newRole.trim();
+				if(newRole.isEmpty() || newRole.equals(null) || newRole.equals("") || newRole.length()==0){
+					String addError = "Invalid. Empty input";
+					request.setAttribute("addError", addError);
 				}
-				if(roleService.checkRoleNameIfExist(newRole)==true){
-					message = " Role already exist. Try another role";
-					request.setAttribute("errors", message);
+				else if(roleService.checkRoleNameIfExist(newRole)==true){
+					String addError = " Role already exist. Try another role";
+					request.setAttribute("addError", addError);
 				}
 				else{
 					roleService.addRole(newRole);
@@ -58,29 +57,45 @@ public class ManageRoles extends HttpServlet{
 	
 		if(action.equals("updateRole")){
 			long updateIdRole = Long.parseLong(request.getParameter("updateIdRole"));
-			String updateRoleName = request.getParameter("updateRoleName");
-			String message= null;
-			if(updateRoleName.isEmpty()){
-				message = "YOu entered an empty field";
-				request.setAttribute("errors", message);
+			String updateRoleName = request.getParameter("updateRoleName").trim();
+			if(updateRoleName.isEmpty() || updateRoleName.equals(null) || updateRoleName.equals("") || updateRoleName.length()==0){
+				String updateError = "Invalid. Empty input";
+				request.setAttribute("updateError", updateError);
 			}
-			if(roleService.checkRoleNameIfExist(updateRoleName)==true){
-				message = " Role already exist. Try another role";
-				request.setAttribute("errors", message);
+			else if(roleService.checkRoleNameIfExist(updateRoleName)==true){
+				String updateError = " Role already exist. Try another role";
+				request.setAttribute("updateError", updateError);
 			}
 			else{
-				roleService.updateRole(updateIdRole, updateRoleName);
+				try{
+					roleService.updateRole(updateIdRole, updateRoleName);
+				}
+				catch (ConstraintViolationException e){
+					String updateError = " Role in use. Cannot delete";
+					request.setAttribute("updateError", updateError);
+				}
 			}
 			
 		}
 		
 		if(action.equals("deleteRole")){
 			long deleteIdRole = Long.parseLong(request.getParameter("deleteIdRole"));
-			roleService.deleteRole(deleteIdRole);
+			Role role = roleService.getRole(deleteIdRole);
+			List<Person> persons = personService.getPersons("id");
+			boolean exist = false;
+			for(Person p : persons){
+				if(p.getRoles().contains(role)==true){
+					String deleteError = " Role in use. Cannot delete";
+					request.setAttribute("deleteError", deleteError);
+					exist = true;
+					break;
+				}
+			}
+			if(exist==false){
+				roleService.deleteRole(deleteIdRole);
+			}
 		}
-		response.sendRedirect(request.getContextPath()+"/ManageRoles");
-		
+		request.setAttribute("roles", roleService.getRoles());
+		request.getRequestDispatcher("/WEB-INF/ViewRoles.jsp").forward(request, response);
 	}
-	
-	
 }
